requirements.txt
langchain>=0.1.0
langchain-community>=0.0.16
langchain-core>=0.1.14
langchain-text-splitters>=0.0.1
ollama>=0.1.0
sentence-transformers>=2.2.2
psycopg2-binary>=2.9.5
pgvector>=0.2.0
click>=8.1.3
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0